+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * `Style/TrivialAccessors` allows predicate methods by default. ([@alexdowad][])
 * [#2869](https://github.com/bbatsov/rubocop/issues/2869): Offenses which occur in the body of a `when` clause with multiple arguments will not be missed. ([@alexdowad][])
 * `Lint/UselessAccessModifier` recognizes method defs inside a `begin` block. ([@alexdowad][])
+* [#2870](https://github.com/bbatsov/rubocop/issues/2870): `Lint/UselessAccessModifier` recognizes method definitions which are passed as an argument to a method call. ([@alexdowad][])
 
 ## 0.37.2 (11/02/2016)